requirements.txt
streamlit
pandas
PyMuPDF
docx
pillow 
python-docx
PyPDF2
PdfReader
tabula
lida
scikit-learn
pandasai   
reportlab
librosa
numpy
opencv-python
tabula-py
google.generativeai
pypdf
transformers
mitosheet
streamlit_shadcn_ui 
scipy
pygwalker
fpdf
matplotlib
firebase_admin
streamlit-cookies-manager
streamlit-lottie
requests